Refactor Code #2: Eliminate in-class iterators
- I used to store iterators in PolyBezierScene and PolyBezierCurve
classes. This has two drawbacks: first, having pointers stored in the
class wouldn't be convenient for me to use STL vector to store the list
of that class. Second, iterators are meant to be used outside of the
class. Storing it inside can lead to misunderstanding and misusing the
iterator.

Note: In PolyBezierCurve, when about to use the iterator, don't forget
to reset the segment index too. For example:

	VectorX2s_iterator it = beginSegments();
	resetToFirstSegment(it);

or:

	VectorX2s_iterator it = endSegments();
	resetToEndSegment(it);